Improve keyboard navigation for Jukebox empty state
Updated the arrow key navigation logic in `handlePlaylistKeyDown` to
target all visible and enabled buttons instead of just elements with
the `.playlist-btn` class. This ensures that keyboard users can easily
navigate the empty state fallback buttons (e.g., "Add Songs" or "Close")
when the playlist is empty, improving overall accessibility.
